requirements.txt
torch~=2.1.0+cu121
torchvision~=0.16.0
numpy~=1.26.0
pillow~=10.0.1
matplotlib~=3.8.0
scikit-learn~=1.2.2